Bug fixes for python.assign() to handle nested "hashes" from R
The following simplified example now works:

    python.assign("foo", list( "key" = c( "subkey" = 13, "anotherkey" = 42 ) ));

A bug in RJSONIO causes toJSON() not to fully-respect the collapse
parameter, so we still do sometimes have newlines despite setting
collapse to the empty string [1].  Using triple single quotes in Python
works around the issue and is more robust anyhow.  Prefixing the Python
string with "r" means that literal escape codes in the JSON itself won't
be interpreted first by Python.  Both of these issues were addressed in
python.call() in previous commits.

The auto-flattening behaviour is also adjusted to only apply to lists.
len(hash) will equal 1 with only a single key in hash, but it shouldn't
be flattened.

[1] Patch: https://github.com/tsibley/RJSONIO/commit/c9196f8 and
    Pull request: https://github.com/duncantl/RJSONIO/pull/15